+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 - **[UPDATE]** Update `Itinerary` to handle segments and connections
 - **[NEW]** Add `ConnectionIcon`
-  [...]
-  
+
 # v40.7.0 (01/10/2020)
 
 - **[FIX]** Fix `statusInformation` position in `TripCard`